feat: move convert to setup
